working modularization
